api/requirements.txt
einops
flask
flask_cors
python-socketio>=5.12.1
flask_socketio
monai>=1.4.0
nibabel>=5.3.2
numpy>=1.26.4
scipy>=1.11.3
torch>=2.6.0
torchvision>=0.18.0
torchaudio
werkzeug
python-dotenv
jwt